baitap-submit/dan-dinh/10-weavite-ui/requirements.txt
weaviate-client==4.10.4
gradio==5.12.0
kagglehub==0.3.6
pandas==2.2.3